refactor(sandbox-ipc): split bridge.ts into spawn.ts + result-classify.ts
Drops bridge.ts under the 400-line soft limit and brings two helper
functions back under the 50-line cap. Extracted modules carry the
process-spawn helpers and worker-result classification logic.
