+++ packages/sandbox/sandbox-ipc/src/bridge.ts
@@ -1,18 +1,22 @@
 import { unlink } from "node:fs/promises";
 import { tmpdir } from "node:os";
 import { join } from "node:path";
-import { serialize } from "node:v8";
 import type { ExecutionContext, JsonObject, Result, SandboxProfile } from "@koi/core";
 import { parseWorkerMessage } from "./protocol.js";
+import {
+  classifyExitCode,
+  createIpcError,
+  mapWorkerErrorCode,
+  processResultMessage,
+} from "./result-classify.js";
+import { buildScrubbedEnv, DEFAULT_ENV_ALLOWLIST, defaultSpawnFn } from "./spawn.js";
 import type {
   BridgeConfig,
   BridgeExecOptions,
   BridgeResult,
-  ErrorMessage,
   IpcError,
   IpcErrorCode,
   IpcProcess,
-  ResultMessage,
   SandboxBridge,
   SpawnFn,
 } from "./types.js";
@@ -23,76 +27,8 @@ const DEFAULT_MAX_RESULT_BYTES = 10_485_760;
 const DEFAULT_SERIALIZATION = "advanced" as const;
 const DEFAULT_TIMEOUT_MS = 30_000;
 
-const SIGNAL_NAMES: Readonly<Record<number, string>> = {
-  1: "SIGHUP",
-  2: "SIGINT",
-  3: "SIGQUIT",
-  6: "SIGABRT",
-  9: "SIGKILL",
-  11: "SIGSEGV",
-  13: "SIGPIPE",
-  14: "SIGALRM",
-  15: "SIGTERM",
-};
-
 export interface CreateSandboxBridgeOptions {
   readonly spawnFn?: SpawnFn;
-}
-
-function createIpcError(
-  code: IpcErrorCode,
-  message: string,
-  details?: {
-    readonly exitCode?: number;
-    readonly signal?: string;
-    readonly durationMs?: number;
-  },
-): IpcError {
-  return {
-    code,
-    message,
-    ...(details?.exitCode !== undefined ? { exitCode: details.exitCode } : {}),
-    ...(details?.signal !== undefined ? { signal: details.signal } : {}),
-    ...(details?.durationMs !== undefined ? { durationMs: details.durationMs } : {}),
-  };
-}
-
-function signalNameFromNumber(signal: number | undefined): string | undefined {
-  if (signal === undefined) {
-    return undefined;
-  }
-  return SIGNAL_NAMES[signal] ?? `SIG${signal}`;
-}
-
-function classifyExitCode(exitCode: number, durationMs: number): IpcError {
-  if (exitCode === 137) {
-    return createIpcError("OOM", "Worker killed by SIGKILL (likely OOM)", {
-      exitCode,
-      signal: "SIGKILL",
-      durationMs,
-    });
-  }
-
-  if (exitCode === 124) {
-    return createIpcError("TIMEOUT", "Worker self-terminated due to timeout", {
-      exitCode,
-      durationMs,
-    });
-  }
-
-  if (exitCode !== 0) {
-    const signal = exitCode > 128 ? signalNameFromNumber(exitCode - 128) : undefined;
-    return createIpcError("CRASH", `Worker exited with code ${exitCode}`, {
-      exitCode,
-      ...(signal !== undefined ? { signal } : {}),
-      durationMs,
-    });
-  }
-
-  return createIpcError("CRASH", "Worker exited cleanly without sending a terminal message", {
-    exitCode,
-    durationMs,
-  });
 }
 
 // BridgeConfig.profile is the enforcement ceiling. Per-call ExecutionContext can only
@@ -148,194 +84,6 @@ function ensureReadablePath(profile: SandboxProfile, path: string): SandboxProfi
     filesystem: {
       ...profile.filesystem,
       allowRead: [...allowRead, path],
-    },
-  };
-}
-
-function mapWorkerErrorCode(code: ErrorMessage["code"]): IpcErrorCode {
-  switch (code) {
-    case "TIMEOUT":
-      return "TIMEOUT";
-    case "OOM":
-      return "OOM";
-    case "PERMISSION":
-      return "PERMISSION";
-    case "CRASH":
-      return "WORKER_ERROR";
-  }
-}
-
-function processResultMessage(
-  message: ResultMessage,
-  maxResultBytes: number,
-  spawnDurationMs: number,
-  serialization: "advanced" | "json",
-): Result<BridgeResult, IpcError> {
-  let sizeBytes: number;
-  try {
-    if (serialization === "advanced") {
-      sizeBytes = serialize(message.output).byteLength;
-    } else {
-      const serialized = JSON.stringify(message.output ?? null);
-      sizeBytes = Buffer.byteLength(serialized, "utf8");
-    }
-  } catch (error) {
-    return {
-      ok: false,
-      error: createIpcError(
-        "DESERIALIZE",
-        `Result could not be serialized for ${serialization} IPC: ${error instanceof Error ? error.message : String(error)}`,
-        { durationMs: message.durationMs },
-      ),
-    };
-  }
-
-  if (sizeBytes > maxResultBytes) {
-    return {
-      ok: false,
-      error: createIpcError(
-        "RESULT_TOO_LARGE",
-        `Result size ${sizeBytes} bytes exceeds limit of ${maxResultBytes} bytes`,
-        { durationMs: message.durationMs },
-      ),
-    };
-  }
-
-  return {
-    ok: true,
-    value: {
-      output: message.output,
-      durationMs: message.durationMs,
-      ...(message.memoryUsedBytes !== undefined
-        ? { memoryUsedBytes: message.memoryUsedBytes }
-        : {}),
-      exitCode: 0,
-      spawnDurationMs,
-    },
-  };
-}
-
-function drainStream(stream: ReadableStream<Uint8Array> | null): void {
-  if (stream === null) {
-    return;
-  }
-
-  void stream
-    .pipeTo(
-      new WritableStream<Uint8Array>({
-        write() {
-          // Discard child stdio without buffering it in host memory.
-        },
-      }),
-    )
-    .catch(() => {});
-}
-
-let detectedSetsidPath: string | null | undefined;
-
-function detectSetsid(): string | null {
-  if (detectedSetsidPath !== undefined) {
-    return detectedSetsidPath;
-  }
-  if (process.platform !== "linux" && process.platform !== "darwin") {
-    detectedSetsidPath = null;
-    return null;
-  }
-  try {
-    detectedSetsidPath = Bun.which("setsid");
-  } catch {
-    detectedSetsidPath = null;
-  }
-  return detectedSetsidPath ?? null;
-}
-
-const DEFAULT_ENV_ALLOWLIST: readonly string[] = [
-  "PATH",
-  "HOME",
-  "USER",
-  "TMPDIR",
-  "LANG",
-  "LC_ALL",
-];
-
-function buildScrubbedEnv(allowlist: readonly string[]): Record<string, string> {
-  const scrubbed: Record<string, string> = {};
-  for (const key of allowlist) {
-    const value = process.env[key];
-    if (typeof value === "string") {
-      scrubbed[key] = value;
-    }
-  }
-  return scrubbed;
-}
-
-function defaultSpawnFn(
-  command: readonly string[],
-  options: {
-    readonly serialization: "advanced" | "json";
-    readonly env?: Readonly<Record<string, string>>;
-    readonly processGroupIsolation?: "required" | "best-effort";
-  },
-): IpcProcess {
-  const messageHandlers: Array<(message: unknown) => void> = [];
-  const setsidPath = detectSetsid();
-  const isolationPolicy = options.processGroupIsolation ?? "required";
-  if (setsidPath === null && isolationPolicy === "required") {
-    throw new Error(
-      "sandbox-ipc: process-group isolation is required but `setsid` is not available on this host. " +
-        'Install `setsid` (util-linux) or set BridgeConfig.processGroupIsolation = "best-effort" to opt out.',
-    );
-  }
-  const useGroup = setsidPath !== null;
-  const spawnArgv: string[] = useGroup ? [setsidPath, "-w", ...command] : [...command];
-
-  const proc = Bun.spawn(spawnArgv, {
-    stdin: "ignore",
-    stdout: "pipe",
-    stderr: "pipe",
-    serialization: options.serialization,
-    env: { ...(options.env ?? {}) },
-    ipc(message: unknown) {
-      for (const handler of messageHandlers) {
-        handler(message);
-      }
-    },
-  });
-
-  drainStream(proc.stdout);
-  drainStream(proc.stderr);
-
-  function killGroup(signalName: NodeJS.Signals): void {
-    if (useGroup) {
-      try {
-        // Negative pid signals the entire process group, terminating any
-        // descendants the worker spawned.
-        process.kill(-proc.pid, signalName);
-        return;
-      } catch {
-        // Fallthrough to direct kill if the group is already gone.
-      }
-    }
-    proc.kill(signalName);
-  }
-
-  return {
-    pid: proc.pid,
-    exited: proc.exited,
-    kill(signal?: number) {
-      const signalName = (signalNameFromNumber(signal ?? 15) ?? "SIGTERM") as NodeJS.Signals;
-      killGroup(signalName);
-    },
-    send(message: unknown) {
-      proc.send(message);
-    },
-    onMessage(handler: (message: unknown) => void) {
-      messageHandlers.push(handler);
-    },
-    onExit(handler: (code: number) => void) {
-      void proc.exited.then((code) => {
-        handler(code);
-      });
     },
   };
 }
